Extract order commercial editor state
